requirements.txt
django==5.0.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
django-unfold==0.20.4
